# Optimized Type Checker performance in UnicodeScalarExtensions and SyntaxText