## Description

This PR optimizes swift code to accommodate slowness of Type Checker.

## Context

The following two methods are currently slow to compile in `swift-syntax` package:

1. `UnicodeScalarExtensions.isValidIdentifierContinuationCodePoint`
2. `SyntaxText.isSlice(of:)`

The optimization is about making the code a bit less "sweet" (as from "syntax sugar"), but a bit faster to get parsed.
Approach is simple: add type annotations, introduce local variables annotated with types.

## Measurements

I have measured the performance of the following command:
```bash
xcodebuild -scheme swift-syntax-Package ONLY_ACTIVE_ARCH=YES -destination 'platform=macos' clean build > /dev/null
```

using `time`. The output of `time` was the following:

Before|After
:-:|:-:
2.10s user 0.53s system 20% cpu 12.784 total|2.08s user 0.51s system 20% cpu 12.629 total
2.11s user 0.50s system 20% cpu 12.896 total|2.09s user 0.50s system 20% cpu 12.802 total
2.12s user 0.50s system 20% cpu 12.720 total|2.07s user 0.52s system 20% cpu 12.793 total
2.10s user 0.51s system 20% cpu 12.905 total|2.08s user 0.53s system 20% cpu 12.768 total
2.05s user 0.47s system 20% cpu 12.638 total|2.07s user 0.53s system 20% cpu 12.710 total
2.12s user 0.52s system 20% cpu 12.569 total|2.05s user 0.50s system 19% cpu 12.846 total
2.14s user 0.53s system 20% cpu 12.792 total|2.07s user 0.50s system 20% cpu 12.514 total


### More Details

I have used https://github.com/qonto/SwiftCompilationTimingParser to measure slow compiling code. The following were the top candidates for optimization:

location|symbol|ms
-- | -- | --
Sources\/SwiftSyntax\/generated\/SyntaxVisitor.swift:3438:16 | instance method visitationFunc(for:) | 407.71
Sources\/SwiftSyntax\/SyntaxText.swift:97:15 | instance method isSlice(of:) | 299.29
Sources\/SwiftSyntax\/generated\/ChildNameForKeyPath.swift:18:13 | global function childName(_:) | 293.93
Sources\/SwiftSyntax\/generated\/SyntaxRewriter.swift:2113:16 | instance method visitationFunc(for:) | 269.33
Sources\/SwiftParser\/Lexer\/UnicodeScalarExtensions.swift:14:52 | getter isValidIdentifierContinuationCodePoint | 123.67
Sources\/SwiftSyntax\/generated\/TokenKind.swift:730:22 | static method fromRaw(kind:text:) | 85.13
Sources\/SwiftSyntax\/generated\/SyntaxEnum.swift:306:8 | instance method as(_:) | 74.07

Out of these, there were only two which are not related to generated code.
After code optimization, the modified methods got the following timing:


location|symbol|ms|improvement
-- | -- | -- | --
UnicodeScalarExtensions.swift:14:52 | getter isValidIdentifierContinuationCodePoint | 20.72 | 83.3%
SyntaxText.swift:97:15 | instance method isSlice(of:) | 7.1 | 97.7%


While the comparison might seem unreasonable to get this PR merged, in my personal opinion and experience, under certain circumstances this type of optimization of swift code for Type Checker might play a big role in overall compilation performance in the long run. What's important is not to introduce a regression, and so thorough measurements should be made.