@art-divin 2023-10-23
👋🏻  @ahoppen ,

> Thanks for investigating swift-syntax’s compile time.
> 
> In this version of the PR code is a lot less readable, especially because of the non-descriptive variable names `first`, `second` etc. Especially in `isSlice(of:)`, I think we should be able to use a lot more descriptive variable names, like `selfEnd` and `otherEnd`, which should improve the code readability and the compile time at the same time.

thank you for your review. I am aware of the "ugliness" of optimized Swift code when it comes to fixing Type Checker performance issues. One of the worst swift code I have ever written was when optimizing Combine closures with explicit return types in a deeply nested context. Though, it reduced compilation duration by 50% in that project.

> For the `isValidIdentifierContinuationCodePoint`, I briefly looked and couldn’t find any descriptive variable names that we could pick. But I think it would be interesting to see if adding `as Bool` could improve the compile time, while leaving the overall shape of the code the same.


I'll do my best and re-run tests with `as Bool` suggestion. I'd also consider adding meaning to the variables. Maybe LLM would be able to come up with such, if no other option would work 😅