@art-divin 2023-10-23
Hey @ahoppen ,

> For the `isValidIdentifierContinuationCodePoint`, I briefly looked and couldn’t find any descriptive variable names that we could pick. But I think it would be interesting to see if adding `as Bool` could improve the compile time, while leaving the overall shape of the code the same.

Indeed, I have applied `as Bool` and it works even faster, just `20.72 ms`! Pushed my commits.