@art-divin 2023-10-23
> Thanks for the quick iteration time @art-divin.
> 
> Could you format the code using swift-format as described in https://github.com/apple/swift-syntax/blob/main/CONTRIBUTING.md#formatting and squash your commits?

Done ✅