@art-divin 2023-10-23
For visibility, I am trying to optimize code in `generated` section of `swift-syntax` as well, by altering the way how code-generator works and adjusting templates. 
My gut feeling is this: 1-2 seconds can be removed (out of 12), but I did not find yet way how to achieve this exactly ⏳